@@ -1,74 +1,21 @@
-__version__ = "1.1.3.dev0"
+from __future__ import annotations
 
+from importlib.metadata import version
+
+from mteb.benchmarks import (
+    MTEB_MAIN_EN,
+    MTEB_RETRIEVAL_LAW,
+    MTEB_RETRIEVAL_WITH_INSTRUCTIONS,
+)
 from mteb.evaluation import *
+from mteb.overview import get_tasks
+
+__version__ = version("mteb")  # fetch version from install metadata
 
 
-MTEB_MAIN_EN = [
-    "AmazonCounterfactualClassification",
-    "AmazonPolarityClassification",
-    "AmazonReviewsClassification",
-    "ArguAna",
-    "ArxivClusteringP2P",
-    "ArxivClusteringS2S",
-    "AskUbuntuDupQuestions",
-    "BIOSSES",
-    "Banking77Classification",
-    "BiorxivClusteringP2P",
-    "BiorxivClusteringS2S",
-    "CQADupstackAndroidRetrieval",
-    "CQADupstackEnglishRetrieval",
-    "CQADupstackGamingRetrieval",
-    "CQADupstackGisRetrieval",
-    "CQADupstackMathematicaRetrieval",
-    "CQADupstackPhysicsRetrieval",
-    "CQADupstackProgrammersRetrieval",
-    "CQADupstackStatsRetrieval",
-    "CQADupstackTexRetrieval",
-    "CQADupstackUnixRetrieval",
-    "CQADupstackWebmastersRetrieval",
-    "CQADupstackWordpressRetrieval",
-    "ClimateFEVER",
-    "DBPedia",
-    "EmotionClassification",
-    "FEVER",
-    "FiQA2018",
-    "HotpotQA",
-    "ImdbClassification",
-    "MSMARCO",
-    "MTOPDomainClassification",
-    "MTOPIntentClassification",
-    "MassiveIntentClassification",
-    "MassiveScenarioClassification",
-    "MedrxivClusteringP2P",
-    "MedrxivClusteringS2S",
-    "MindSmallReranking",
-    "NFCorpus",
-    "NQ",
-    "QuoraRetrieval",
-    "RedditClustering",
-    "RedditClusteringP2P",
-    "SCIDOCS",
-    "SICK-R",
-    "STS12",
-    "STS13",
-    "STS14",
-    "STS15",
-    "STS16",
-    "STS17",
-    "STS22",
-    "STSBenchmark",
-    "SciDocsRR",
-    "SciFact",
-    "SprintDuplicateQuestions",
-    "StackExchangeClustering",
-    "StackExchangeClusteringP2P",
-    "StackOverflowDupQuestions",
-    "SummEval",
-    "TRECCOVID",
-    "Touche2020",
-    "ToxicConversationsClassification",
-    "TweetSentimentExtractionClassification",
-    "TwentyNewsgroupsClustering",
-    "TwitterSemEval2015",
-    "TwitterURLCorpus"
+__all__ = [
+    "MTEB_MAIN_EN",
+    "MTEB_RETRIEVAL_LAW",
+    "MTEB_RETRIEVAL_WITH_INSTRUCTIONS",
+    "get_tasks",
 ]
