@@ -1,3 +1,6 @@
+from __future__ import annotations
+
+import logging
 import random
 from abc import ABC, abstractmethod
 
@@ -5,8 +8,15 @@ import datasets
 import numpy as np
 import torch
 
+from mteb.abstasks.TaskMetadata import TaskMetadata
+
+logger = logging.getLogger(__name__)
+
 
 class AbsTask(ABC):
+    metadata: TaskMetadata
+    superseeded_by: None | str = None
+
     def __init__(self, seed=42, **kwargs):
         self.dataset = None
         self.data_loaded = False
@@ -20,39 +30,88 @@ class AbsTask(ABC):
         torch.manual_seed(self.seed)
         torch.cuda.manual_seed_all(self.seed)
 
+    def check_if_dataset_is_superseeded(self):
+        """Check if the dataset is superseeded by a newer version"""
+        if self.superseeded_by:
+            logger.warning(
+                f"Dataset '{self.metadata.name}' is superseeded by '{self.superseeded_by}', you might consider using the newer version of the dataset."
+            )
+
+    def dataset_transform(self):
+        """Transform operations applied to the dataset after loading.
+        Override this method if your dataset requires any transformation.
+        """
+        pass
+
+    @staticmethod
+    def stratified_subsampling(
+        dataset_dict: datasets.DatasetDict,
+        seed: int,
+        splits: list[str] = ["test"],
+        label: str = "label",
+        n_samples: int = 2048,
+    ) -> datasets.DatasetDict:
+        """Subsamples the dataset with stratification by the supplied label.
+        Returns a datasetDict object.
+
+        Args:
+            dataset_dict: the DatasetDict object.
+            seed: the random seed.
+            splits: the splits of the dataset.
+            label: the label with which the stratified sampling is based on.
+            n_samples: Optional, number of samples to subsample. Default is max_n_samples.
+        """
+        ## Can only do this if the label column is of ClassLabel.
+        if not isinstance(dataset_dict[splits[0]].features[label], datasets.ClassLabel):
+            dataset_dict = dataset_dict.class_encode_column(label)
+
+        for split in splits:
+            dataset_dict.update(
+                {
+                    split: dataset_dict[split].train_test_split(
+                        test_size=n_samples, seed=seed, stratify_by_column=label
+                    )["test"]
+                }
+            )  ## only take the specified test split.
+        return dataset_dict
+
     def load_data(self, **kwargs):
-        """
-        Load dataset from HuggingFace hub
-        """
+        """Load dataset from HuggingFace hub"""
         if self.data_loaded:
             return
-
-        # TODO: add split argument
-        self.dataset = datasets.load_dataset(
-            self.description["hf_hub_name"], revision=self.description.get("revision", None)
-        )
+        self.dataset = datasets.load_dataset(**self.metadata_dict["dataset"])
+        self.dataset_transform()
         self.data_loaded = True
 
     @property
-    @abstractmethod
-    def description(self):
-        """
-        Returns a description of the task. Should contain the following fields:
-        name: Name of the task (usually equal to the class name. Should be a valid name for a path on disc)
-        description: Longer description & references for the task
-        type: Of the set: [sts]
-        eval_splits: Splits used for evaluation as list, e.g. ['dev', 'test']
-        main_score: Main score value for task
-        """
-        raise NotImplementedError
+    def metadata_dict(self) -> dict[str, str]:
+        metadata_dict = dict(self.metadata)
+        return metadata_dict
 
     @abstractmethod
     def evaluate(self, model, split="test"):
-        """
-        Evaluates a Sentence Embedding Model on the task.
+        """Evaluates a Sentence Embedding Model on the task.
         Returns a dict (that can be serialized to json).
         :param model: Sentence embedding method. Implements a encode(sentences) method, that encodes sentences
         and returns a numpy matrix with the sentence embeddings
         :param split: Which datasplit to be used.
         """
         raise NotImplementedError
+
+    @property
+    def languages(self) -> list[str]:
+        """Returns the languages of the task"""
+        return self.metadata.languages
+
+    def __repr__(self) -> str:
+        """Format the representation of the task such that it appears as:
+
+        TaskObjectName(name='{name}', languages={lang1, lang2, ...})
+        """
+        langs = self.languages
+        if len(langs) > 3:
+            langs = langs[:3]
+            langs.append("...")
+        return (
+            f"{self.__class__.__name__}(name='{self.metadata.name}', languages={langs})"
+        )
