@@ -1,15 +1,15 @@
+from __future__ import annotations
+
 from .AbsTask import *
 from .AbsTaskBitextMining import *
 from .AbsTaskClassification import *
 from .AbsTaskClustering import *
+from .AbsTaskInstructionRetrieval import *
 from .AbsTaskPairClassification import *
 from .AbsTaskReranking import *
 from .AbsTaskRetrieval import *
 from .AbsTaskSTS import *
 from .AbsTaskSummarization import *
-from .BeIRKOTask import *
-from .BeIRPLTask import *
-from .BeIRTask import *
 from .CrosslingualTask import *
 from .LangMapping import *
 from .MultilingualTask import *
