@@ -1,257 +1,498 @@
-import logging
-import inspect
-from typing import Callable, Dict, List, Set
+from __future__ import annotations
 
-import numpy as np
+import heapq
+import json
+import logging
+import os
+from collections import defaultdict
+from typing import Dict, List, Tuple
+
+import pytrec_eval
 import torch
-from tqdm import trange
+import tqdm
+from sentence_transformers import CrossEncoder, SentenceTransformer
+from sentence_transformers.models import Transformer, WordEmbeddings
 
 from .Evaluator import Evaluator
-from .utils import cos_sim, dot_score
+from .utils import cos_sim, dot_score, download, hole, mrr, recall_cap, top_k_accuracy
 
 logger = logging.getLogger(__name__)
 
 
-class RetrievalEvaluator(Evaluator):
-    """
-    This class evaluates an Information Retrieval (IR) setting.
-    Given a set of queries and a large corpus set. It will retrieve for each query the top-k most similar document. It measures
-    Mean Reciprocal Rank (MRR), Recall@k, and Normalized Discounted Cumulative Gain (NDCG)
-    """
-
+# Adapted from https://github.com/beir-cellar/beir/blob/f062f038c4bfd19a8ca942a9910b1e0d218759d4/beir/retrieval/search/dense/exact_search.py#L12
+class DenseRetrievalExactSearch:
     def __init__(
         self,
-        queries: Dict[str, str],  # qid => query
-        corpus: Dict[str, str],  # cid => doc
-        relevant_docs: Dict[str, Set[str]],  # qid => Set[cid]
+        model,
+        batch_size: int = 128,
         corpus_chunk_size: int = 50000,
-        mrr_at_k: List[int] = [10],
-        ndcg_at_k: List[int] = [10],
-        accuracy_at_k: List[int] = [1, 3, 5, 10],
-        precision_recall_at_k: List[int] = [1, 3, 5, 10],
-        map_at_k: List[int] = [100],
-        show_progress_bar: bool = False,
-        batch_size: int = 32,
-        name: str = "",
-        score_functions: List[Callable[[torch.Tensor, torch.Tensor], torch.Tensor]] = {
-            "cos_sim": cos_sim,
-            "dot": dot_score,
-        },  # Score function, higher=more similar
-        main_score_function: str = None,
-        limit: int = None,
-        **kwargs
+        previous_results: str = None,
+        **kwargs,
     ):
-        super().__init__(**kwargs)
-        self.queries_ids = []
-        for qid in queries:
-            if qid in relevant_docs and len(relevant_docs[qid]) > 0:
-                self.queries_ids.append(qid)
-                if limit and len(self.queries_ids) >= limit:
-                    break
-
-        self.queries = [queries[qid] for qid in self.queries_ids]
-
-        self.corpus_ids = list(corpus.keys())
-        self.corpus = [corpus[cid] for cid in self.corpus_ids]
-
-        self.relevant_docs = relevant_docs
-        self.corpus_chunk_size = corpus_chunk_size
-        self.mrr_at_k = mrr_at_k
-        self.ndcg_at_k = ndcg_at_k
-        self.accuracy_at_k = accuracy_at_k
-        self.precision_recall_at_k = precision_recall_at_k
-        self.map_at_k = map_at_k
-
-        self.show_progress_bar = show_progress_bar
+        # Model is class that provides encode_corpus() and encode_queries()
+        self.model = model
         self.batch_size = batch_size
-        self.name = name
-        self.score_functions = score_functions
-        self.score_function_names = sorted(list(self.score_functions.keys()))
-        self.main_score_function = main_score_function
+        self.score_functions = {"cos_sim": cos_sim, "dot": dot_score}
+        self.score_function_desc = {
+            "cos_sim": "Cosine Similarity",
+            "dot": "Dot Product",
+        }
+        self.corpus_chunk_size = corpus_chunk_size
+        self.previous_results = previous_results
+        self.show_progress_bar = kwargs.get("show_progress_bar", True)
+        self.convert_to_tensor = kwargs.get("convert_to_tensor", True)
+        self.save_corpus_embeddings = kwargs.get("save_corpus_embeddings", False)
+        self.corpus_embeddings = defaultdict(list)
+        self.results = {}
 
-    def __call__(self, model) -> float:
-        scores = self.compute_metrics(model)
+        if self.previous_results is not None:
+            self.previous_results = self.load_results_file()
 
-        if self.main_score_function is None:
-            scores["main_score"] = max(
-                [scores[name]["map_at_" + str(max(self.map_at_k))] for name in self.score_function_names]
+        if isinstance(self.model, CrossEncoder):
+            # load the predict instance from the CrossEncoder
+            # custom functions can be used by extending the DenseRetrievalExactSearch class
+            self.predict = self.model.predict
+
+    def search(
+        self,
+        corpus: dict[str, dict[str, str]],
+        queries: dict[str, str],
+        top_k: int,
+        score_function: str,
+        return_sorted: bool = False,
+        **kwargs,
+    ) -> dict[str, dict[str, float]]:
+        # Create embeddings for all queries using model.encode_queries()
+        # Runs semantic search against the corpus embeddings
+        # Returns a ranked list with the corpus ids
+        if score_function not in self.score_functions:
+            raise ValueError(
+                "score function: {} must be either (cos_sim) for cosine similarity or (dot) for dot product".format(
+                    score_function
+                )
             )
-        else:
-            scores["main_score"] = scores[self.main_score_function]["map_at_" + str(max(self.map_at_k))]
-        return scores
 
-    def compute_metrics(self, model, corpus_model=None, corpus_embeddings: torch.Tensor = None) -> Dict[str, float]:
-        if corpus_model is None:
-            corpus_model = model
-
-        max_k = max(
-            max(self.mrr_at_k),
-            max(self.ndcg_at_k),
-            max(self.accuracy_at_k),
-            max(self.precision_recall_at_k),
-            max(self.map_at_k),
+        logger.info("Encoding Queries...")
+        query_ids = list(queries.keys())
+        self.results = {qid: {} for qid in query_ids}
+        queries = [queries[qid] for qid in queries]
+        query_embeddings = self.model.encode_queries(
+            queries,
+            batch_size=self.batch_size,
+            show_progress_bar=self.show_progress_bar,
+            convert_to_tensor=self.convert_to_tensor,
+            **kwargs,
         )
 
-        # Compute embedding for the queries
-        logger.info("Encoding the queries...")
-        # We don't know if encode has the kwargs show_progress_bar
-        kwargs = {
-            "show_progress_bar": self.show_progress_bar
-        } if "show_progress_bar" in inspect.signature(model.encode).parameters else {}
-        query_embeddings = np.asarray(model.encode(self.queries, batch_size=self.batch_size, **kwargs))
-        queries_result_list = {}
-        for name in self.score_functions:
-            queries_result_list[name] = [[] for _ in range(len(query_embeddings))]
+        logger.info("Sorting Corpus by document length (Longest first)...")
+        corpus_ids = sorted(
+            corpus,
+            key=lambda k: len(corpus[k].get("title", "") + corpus[k].get("text", "")),
+            reverse=True,
+        )
+        corpus = [corpus[cid] for cid in corpus_ids]
 
-        # Iterate over chunks of the corpus
-        logger.info("Encoding chunks of corpus, and computing similarity scores with queries...")
-        for corpus_start_idx in trange(
-            0,
-            len(self.corpus),
-            self.corpus_chunk_size,
-            desc="Corpus Chunks",
-            disable=not self.show_progress_bar,
-        ):
+        logger.info("Encoding Corpus in batches... Warning: This might take a while!")
+        logger.info(
+            "Scoring Function: {} ({})".format(
+                self.score_function_desc[score_function], score_function
+            )
+        )
+
+        itr = range(0, len(corpus), self.corpus_chunk_size)
+
+        result_heaps = {
+            qid: [] for qid in query_ids
+        }  # Keep only the top-k docs for each query
+        for batch_num, corpus_start_idx in enumerate(itr):
+            logger.info("Encoding Batch {}/{}...".format(batch_num + 1, len(itr)))
+            corpus_end_idx = min(corpus_start_idx + self.corpus_chunk_size, len(corpus))
+
             # Encode chunk of corpus
-            if corpus_embeddings is None:
-                corpus_end_idx = min(corpus_start_idx + self.corpus_chunk_size, len(self.corpus))
-                sub_corpus_embeddings = np.asarray(corpus_model.encode(
-                    self.corpus[corpus_start_idx:corpus_end_idx],
-                    batch_size=self.batch_size,
-                ))
+            if (
+                self.save_corpus_embeddings
+                and "qid" in kwargs
+                and len(self.corpus_embeddings[kwargs["qid"]])
+            ):
+                sub_corpus_embeddings = torch.tensor(
+                    self.corpus_embeddings[kwargs["qid"]][batch_num]
+                )
             else:
-                corpus_end_idx = min(corpus_start_idx + self.corpus_chunk_size, len(corpus_embeddings))
-                sub_corpus_embeddings = corpus_embeddings[corpus_start_idx:corpus_end_idx]
-
-            # Compute cosine similarites
-            for name, score_function in self.score_functions.items():
-                pair_scores = score_function(query_embeddings, sub_corpus_embeddings)
-
-                # Get top-k values
-                pair_scores_top_k_values, pair_scores_top_k_idx = torch.topk(
-                    pair_scores,
-                    min(max_k, len(pair_scores[0])),
-                    dim=1,
-                    largest=True,
-                    sorted=False,
+                # Encode chunk of corpus
+                sub_corpus_embeddings = self.model.encode_corpus(
+                    corpus[corpus_start_idx:corpus_end_idx],
+                    batch_size=self.batch_size,
+                    show_progress_bar=self.show_progress_bar,
+                    convert_to_tensor=self.convert_to_tensor,
                 )
-                pair_scores_top_k_values = pair_scores_top_k_values.cpu().tolist()
-                pair_scores_top_k_idx = pair_scores_top_k_idx.cpu().tolist()
+                if self.save_corpus_embeddings and "qid" in kwargs:
+                    self.corpus_embeddings[kwargs["qid"]].append(sub_corpus_embeddings)
 
-                for query_itr in range(len(query_embeddings)):
-                    for sub_corpus_id, score in zip(
-                        pair_scores_top_k_idx[query_itr],
-                        pair_scores_top_k_values[query_itr],
-                    ):
-                        corpus_id = self.corpus_ids[corpus_start_idx + sub_corpus_id]
-                        queries_result_list[name][query_itr].append({"corpus_id": corpus_id, "score": score})
+            # Compute similarites using either cosine-similarity or dot product
+            cos_scores = self.score_functions[score_function](
+                query_embeddings, sub_corpus_embeddings
+            )
+            cos_scores[torch.isnan(cos_scores)] = -1
 
-        # Compute scores
-        logger.info("Computing metrics...")
-        scores = {name: self._compute_metrics(queries_result_list[name]) for name in self.score_functions}
+            # Get top-k values
+            cos_scores_top_k_values, cos_scores_top_k_idx = torch.topk(
+                cos_scores,
+                min(
+                    top_k + 1,
+                    len(cos_scores[1]) if len(cos_scores) > 1 else len(cos_scores[-1]),
+                ),
+                dim=1,
+                largest=True,
+                sorted=return_sorted,
+            )
+            cos_scores_top_k_values = cos_scores_top_k_values.cpu().tolist()
+            cos_scores_top_k_idx = cos_scores_top_k_idx.cpu().tolist()
 
-        return scores
+            for query_itr in range(len(query_embeddings)):
+                query_id = query_ids[query_itr]
+                for sub_corpus_id, score in zip(
+                    cos_scores_top_k_idx[query_itr], cos_scores_top_k_values[query_itr]
+                ):
+                    corpus_id = corpus_ids[corpus_start_idx + sub_corpus_id]
+                    if corpus_id != query_id:
+                        if len(result_heaps[query_id]) < top_k:
+                            # Push item on the heap
+                            heapq.heappush(result_heaps[query_id], (score, corpus_id))
+                        else:
+                            # If item is larger than the smallest in the heap, push it on the heap then pop the smallest element
+                            heapq.heappushpop(
+                                result_heaps[query_id], (score, corpus_id)
+                            )
 
-    def _compute_metrics(self, queries_result_list: List[object]):
+        for qid in result_heaps:
+            for score, corpus_id in result_heaps[qid]:
+                self.results[qid][corpus_id] = score
+
+        return self.results
+
+    def load_results_file(self):
+        # load the first stage results from file in format {qid: {doc_id: score}}
+        if "https://" in self.previous_results:
+            # download the file
+            if not os.path.exists(self.previous_results):
+                url_descriptor = self.previous_results.split("https://")[-1].replace(
+                    "/", "--"
+                )
+                dest_file = os.path.join(
+                    "results", f"cached_predictions--{url_descriptor}"
+                )
+                os.makedirs(os.path.dirname(os.path.abspath(dest_file)), exist_ok=True)
+                download(self.previous_results, dest_file)
+                logger.info(
+                    f"Downloaded the previous results at {self.previous_results} to {dest_file}"
+                )
+            self.previous_results = dest_file
+
+        with open(self.previous_results, "r") as f:
+            previous_results = json.load(f)
+        assert isinstance(previous_results, dict)
+        assert isinstance(previous_results[list(previous_results.keys())[0]], dict)
+        return previous_results
+
+    def search_cross_encoder(
+        self,
+        corpus: Dict[str, Dict[str, str]],
+        queries: Dict[str, str],
+        top_k: int,
+        instructions: Dict[str, str] | None = None,
+        **kwargs,
+    ) -> Dict[str, Dict[str, float]]:
+        """This function provides support for reranker (or cross-encoder) models that encoder query and document at the same time (typically with attention).
+        Some notable examples include MonoBERT, MonoT5, RankLlama, etc.
+        Note: you must provide the path to the results to rerank to the __init__ function as `previous_results`
         """
-        Compute metrics for a list of queries
+        pairs = []  # create the pairs for reranking
+        for qid in queries.keys():
+            q_results = self.previous_results[qid]
+            # take the top-k only
+            q_results_sorted = {
+                k: v
+                for k, v in sorted(
+                    q_results.items(), key=lambda item: item[1], reverse=True
+                )
+            }
+            top_n = [k for k, v in list(q_results_sorted.items())[:top_k]]
+            query = queries[qid]
+            for doc_id in top_n:
+                corpus_item = (
+                    corpus[doc_id].get("title", "") + " " + corpus[doc_id]["text"]
+                ).strip()
+                pairs.append(
+                    (
+                        query,
+                        corpus_item,
+                        instructions[query] if instructions is not None else None,
+                        qid,
+                        doc_id,
+                    )
+                )
 
-        Args:
-            queries_result_list (`List[List[Dict]]`): List of lists of dictionaries with keys "corpus_id" and "score"
+        logger.info(f"Reranking the top {top_k} in batches... This might take a while!")
+        itr = range(0, len(pairs), self.batch_size)
 
-        Returns:
-            `Dict[str, Dict[str, float]]`: Dictionary with keys "mrr@k", "ndcg@k", "accuracy@k", "precision_recall@k", "map@k"
-                which values are dictionaries with scores for different k values
-        """
-        # Init score computation values
-        num_hits_at_k = {"accuracy_at_" + str(k): 0 for k in self.accuracy_at_k}
-        precisions_at_k = {"precision_at_" + str(k): [] for k in self.precision_recall_at_k}
-        recall_at_k = {"recall_at_" + str(k): [] for k in self.precision_recall_at_k}
-        MRR = {"mrr_at_" + str(k): 0 for k in self.mrr_at_k}
-        ndcg = {"ndcg_at_" + str(k): [] for k in self.ndcg_at_k}
-        AveP_at_k = {"map_at_" + str(k): [] for k in self.map_at_k}
+        results = {qid: {} for qid in queries.keys()}
+        for batch_num, corpus_start_idx in enumerate(
+            tqdm.tqdm(itr, leave=False, disable=not self.show_progress_bar)
+        ):
+            corpus_end_idx = min(corpus_start_idx + self.batch_size, len(pairs))
+            cur_batch = pairs[corpus_start_idx:corpus_end_idx]
 
-        # Compute scores on results
-        for query_itr in range(len(queries_result_list)):
-            query_id = self.queries_ids[query_itr]
+            (
+                queries_in_pair,
+                corpus_in_pair,
+                instructions_in_pair,
+                query_ids,
+                corpus_ids,
+            ) = zip(*cur_batch)
 
-            # Sort scores
-            top_hits = sorted(queries_result_list[query_itr], key=lambda x: x["score"], reverse=True)
-            query_relevant_docs = self.relevant_docs[query_id]
+            assert (
+                len(queries_in_pair) == len(corpus_in_pair) == len(instructions_in_pair)
+            )
 
-            # Accuracy@k - We count the result correct, if at least one relevant doc is accross the top-k documents
-            for k_val in self.accuracy_at_k:
-                for hit in top_hits[0:k_val]:
-                    if hit["corpus_id"] in query_relevant_docs:
-                        num_hits_at_k["accuracy_at_" + str(k_val)] += 1
-                        break
-
-            # Precision and Recall@k
-            for k_val in self.precision_recall_at_k:
-                num_correct = 0
-                for hit in top_hits[0:k_val]:
-                    if hit["corpus_id"] in query_relevant_docs:
-                        num_correct += 1
-
-                precisions_at_k["precision_at_" + str(k_val)].append(num_correct / k_val)
-                recall_at_k["recall_at_" + str(k_val)].append(num_correct / len(query_relevant_docs))
-
-            # MRR@k
-            for k_val in self.mrr_at_k:
-                for rank, hit in enumerate(top_hits[0:k_val]):
-                    if hit["corpus_id"] in query_relevant_docs:
-                        MRR["mrr_at_" + str(k_val)] += 1.0 / (rank + 1)
-                        break
-
-            # NDCG@k
-            for k_val in self.ndcg_at_k:
-                predicted_relevance = [
-                    1 if top_hit["corpus_id"] in query_relevant_docs else 0 for top_hit in top_hits[0:k_val]
+            if isinstance(self.model, CrossEncoder):
+                # can't take instructions, so add them here
+                queries_in_pair = [
+                    f"{q} {i}".strip()
+                    for i, q in zip(instructions_in_pair, queries_in_pair)
                 ]
-                true_relevances = [1] * len(query_relevant_docs)
-
-                ndcg_value = self.compute_dcg_at_k(predicted_relevance, k_val) / self.compute_dcg_at_k(
-                    true_relevances, k_val
+                scores = self.model.predict(list(zip(queries_in_pair, corpus_in_pair)))
+            else:
+                # may use the instructions in a unique way, so give them also
+                scores = self.model.predict(
+                    list(zip(queries_in_pair, corpus_in_pair, instructions_in_pair))
                 )
-                ndcg["ndcg_at_" + str(k_val)].append(ndcg_value)
 
-            # MAP@k
-            for k_val in self.map_at_k:
-                num_correct = 0
-                sum_precisions = 0
+            for i, score in enumerate(scores):
+                results[query_ids[i]][corpus_ids[i]] = float(score)
 
-                for rank, hit in enumerate(top_hits[0:k_val]):
-                    if hit["corpus_id"] in query_relevant_docs:
-                        num_correct += 1
-                        sum_precisions += num_correct / (rank + 1)
+        return results
 
-                avg_precision = sum_precisions / min(k_val, len(query_relevant_docs))
-                AveP_at_k["map_at_" + str(k_val)].append(avg_precision)
+    def predict(self, queries, passages, **kwargs):
+        raise NotImplementedError(
+            "You must implement a predict method for your reranker model"
+        )
 
-        # Compute averages
-        for k in num_hits_at_k:
-            num_hits_at_k[k] /= len(self.queries)
 
-        for k in precisions_at_k:
-            precisions_at_k[k] = np.mean(precisions_at_k[k])
+class DRESModel:
+    """Dense Retrieval Exact Search (DRES) requires an encode_queries & encode_corpus method.
+    This class converts a model with just an .encode method into DRES format.
+    """
 
-        for k in recall_at_k:
-            recall_at_k[k] = np.mean(recall_at_k[k])
+    def __init__(self, model, **kwargs):
+        self.model = model
+        self.use_sbert_model = isinstance(model, SentenceTransformer)
+        self.save_corpus_embeddings = kwargs.get("save_corpus_embeddings", False)
+        self.corpus_embeddings = {}
 
-        for k in ndcg:
-            ndcg[k] = np.mean(ndcg[k])
+    def encode_queries(self, queries: List[str], batch_size: int, **kwargs):
+        if self.use_sbert_model:
+            if isinstance(self.model._first_module(), Transformer):
+                logger.info(
+                    f"Queries will be truncated to {self.model.get_max_seq_length()} tokens."
+                )
+            elif isinstance(self.model._first_module(), WordEmbeddings):
+                logger.warning(
+                    "Queries will not be truncated. This could lead to memory issues. In that case please lower the batch_size."
+                )
 
-        for k in MRR:
-            MRR[k] /= len(self.queries)
+        if "instructions" in kwargs:
+            if kwargs["instructions"] is not None:
+                queries = [
+                    (query + " " + kwargs["instructions"][query]).strip()
+                    for query in queries
+                ]
+            new_kwargs = {
+                k: v for k, v in kwargs.items() if k not in ["instructions", "qid"]
+            }
+        else:
+            # can't just delete, cuz assign by reference on kwargs
+            new_kwargs = kwargs
 
-        for k in AveP_at_k:
-            AveP_at_k[k] = np.mean(AveP_at_k[k])
+        return self.model.encode(queries, batch_size=batch_size, **new_kwargs)
 
-        return {**num_hits_at_k, **precisions_at_k, **recall_at_k, **MRR, **ndcg, **AveP_at_k}
+    def encode_corpus(self, corpus: List[Dict[str, str]], batch_size: int, **kwargs):
+        if (
+            "qid" in kwargs
+            and self.save_corpus_embeddings
+            and len(self.corpus_embeddings) > 0
+        ):
+            return self.corpus_embeddings[kwargs["qid"]]
+
+        if isinstance(corpus, dict):
+            sentences = [
+                (corpus["title"][i] + " " + corpus["text"][i]).strip()
+                if "title" in corpus
+                else corpus["text"][i].strip()
+                for i in range(len(corpus["text"]))
+            ]
+        else:
+            sentences = [
+                (doc["title"] + " " + doc["text"]).strip()
+                if "title" in doc
+                else doc["text"].strip()
+                for doc in corpus
+            ]
+
+        if "instructions" in kwargs:  # not used on the doc side
+            new_kwargs = {
+                k: v for k, v in kwargs.items() if k not in ["instructions", "qid"]
+            }
+        else:
+            # can't just delete, cuz assign by reference on kwargs
+            new_kwargs = kwargs
+
+        corpus_embeddings = self.model.encode(
+            sentences, batch_size=batch_size, **new_kwargs
+        )
+        if self.save_corpus_embeddings and "qid" in kwargs:
+            if isinstance(corpus_embeddings, torch.tensor):
+                corpus_embeddings = corpus_embeddings.cpu().detach()
+            self.corpus_embeddings[kwargs["qid"]] = corpus_embeddings
+        return corpus_embeddings
+
+    def encode(self, sentences: List[str], **kwargs):
+        return self.model.encode(sentences, **kwargs)
+
+
+def is_dres_compatible(model):
+    for method in ["encode_queries", "encode_corpus"]:
+        op = getattr(model, method, None)
+        if not (callable(op)):
+            return False
+    return True
+
+
+def is_cross_encoder_compatible(model):
+    op = getattr(model, "predict", None)
+    if not (callable(op)):
+        return False
+    return True
+
+
+# Adapted from https://github.com/beir-cellar/beir/blob/f062f038c4bfd19a8ca942a9910b1e0d218759d4/beir/retrieval/evaluation.py#L9
+class RetrievalEvaluator(Evaluator):
+    def __init__(
+        self,
+        retriever=None,
+        k_values: List[int] = [1, 3, 5, 10, 20, 100, 1000],
+        score_function: str = "cos_sim",
+        **kwargs,
+    ):
+        super().__init__(**kwargs)
+        self.is_cross_encoder = False
+        if is_cross_encoder_compatible(retriever):
+            logger.info(
+                "The custom predict function of the model will be used if not a SentenceTransformer CrossEncoder"
+            )
+            self.retriever = DenseRetrievalExactSearch(retriever, **kwargs)
+            self.is_cross_encoder = True
+        elif is_dres_compatible(retriever):
+            logger.info(
+                "The custom encode_queries and encode_corpus functions of the model will be used"
+            )
+            self.retriever = DenseRetrievalExactSearch(retriever, **kwargs)
+        else:
+            self.retriever = DenseRetrievalExactSearch(DRESModel(retriever), **kwargs)
+        self.k_values = k_values
+        self.top_k = (
+            max(k_values) if "top_k" not in kwargs else kwargs["top_k"]
+        )  # can lower it if reranking
+        self.score_function = score_function
+
+    def __call__(
+        self, corpus: dict[str, dict[str, str]], queries: dict[str, str]
+    ) -> dict[str, dict[str, float]]:
+        if not self.retriever:
+            raise ValueError("Model/Technique has not been provided!")
+
+        if self.is_cross_encoder:
+            return self.retriever.search_cross_encoder(corpus, queries, self.top_k)
+        else:
+            return self.retriever.search(
+                corpus, queries, self.top_k, self.score_function
+            )
 
     @staticmethod
-    def compute_dcg_at_k(relevances, k):
-        dcg = 0
-        for i in range(min(len(relevances), k)):
-            dcg += relevances[i] / np.log2(i + 2)  # +2 as we start our idx at 0
-        return dcg
+    def evaluate(
+        qrels: dict[str, dict[str, int]],
+        results: dict[str, dict[str, float]],
+        k_values: List[int],
+        ignore_identical_ids: bool = True,
+    ) -> Tuple[Dict[str, float], dict[str, float], dict[str, float], dict[str, float]]:
+        if ignore_identical_ids:
+            logger.info(
+                "For evaluation, we ignore identical query and document ids (default), please explicitly set ``ignore_identical_ids=False`` to ignore this."
+            )
+            popped = []
+            for qid, rels in results.items():
+                for pid in list(rels):
+                    if qid == pid:
+                        results[qid].pop(pid)
+                        popped.append(pid)
+
+        ndcg = {}
+        _map = {}
+        recall = {}
+        precision = {}
+
+        for k in k_values:
+            ndcg[f"NDCG@{k}"] = 0.0
+            _map[f"MAP@{k}"] = 0.0
+            recall[f"Recall@{k}"] = 0.0
+            precision[f"P@{k}"] = 0.0
+
+        map_string = "map_cut." + ",".join([str(k) for k in k_values])
+        ndcg_string = "ndcg_cut." + ",".join([str(k) for k in k_values])
+        recall_string = "recall." + ",".join([str(k) for k in k_values])
+        precision_string = "P." + ",".join([str(k) for k in k_values])
+        evaluator = pytrec_eval.RelevanceEvaluator(
+            qrels, {map_string, ndcg_string, recall_string, precision_string}
+        )
+        scores = evaluator.evaluate(results)
+
+        for query_id in scores.keys():
+            for k in k_values:
+                ndcg[f"NDCG@{k}"] += scores[query_id]["ndcg_cut_" + str(k)]
+                _map[f"MAP@{k}"] += scores[query_id]["map_cut_" + str(k)]
+                recall[f"Recall@{k}"] += scores[query_id]["recall_" + str(k)]
+                precision[f"P@{k}"] += scores[query_id]["P_" + str(k)]
+
+        for k in k_values:
+            ndcg[f"NDCG@{k}"] = round(ndcg[f"NDCG@{k}"] / len(scores), 5)
+            _map[f"MAP@{k}"] = round(_map[f"MAP@{k}"] / len(scores), 5)
+            recall[f"Recall@{k}"] = round(recall[f"Recall@{k}"] / len(scores), 5)
+            precision[f"P@{k}"] = round(precision[f"P@{k}"] / len(scores), 5)
+
+        for eval in [ndcg, _map, recall, precision]:
+            logger.info("\n")
+            for k in eval.keys():
+                logger.info("{}: {:.4f}".format(k, eval[k]))
+
+        return ndcg, _map, recall, precision
+
+    @staticmethod
+    def evaluate_custom(
+        qrels: dict[str, dict[str, int]],
+        results: dict[str, dict[str, float]],
+        k_values: List[int],
+        metric: str,
+    ) -> Tuple[Dict[str, float]]:
+        if metric.lower() in ["mrr", "mrr@k", "mrr_cut"]:
+            return mrr(qrels, results, k_values)
+        elif metric.lower() in ["recall_cap", "r_cap", "r_cap@k"]:
+            return recall_cap(qrels, results, k_values)
+        elif metric.lower() in ["hole", "hole@k"]:
+            return hole(qrels, results, k_values)
+        elif metric.lower() in [
+            "acc",
+            "top_k_acc",
+            "accuracy",
+            "accuracy@k",
+            "top_k_accuracy",
+        ]:
+            return top_k_accuracy(qrels, results, k_values)
