@@ -1,11 +1,20 @@
+from __future__ import annotations
+
+import logging
+from typing import Dict, List, Tuple
+
+import pandas as pd
+import requests
 import torch
+import tqdm
 
 
 def cos_sim(a, b):
-    """
-    Computes the cosine similarity cos_sim(a[i], b[j]) for all i and j.
-    :return: Matrix with res[i][j]  = cos_sim(a[i], b[j])
-    """
+    """Computes the cosine similarity cos_sim(a[i], b[j]) for all i and j.
+
+    Return:
+        Matrix with res[i][j]  = cos_sim(a[i], b[j])
+    """  # noqa: D402
     if not isinstance(a, torch.Tensor):
         a = torch.tensor(a)
 
@@ -24,8 +33,7 @@ def cos_sim(a, b):
 
 
 def dot_score(a: torch.Tensor, b: torch.Tensor):
-    """
-    Computes the dot-product dot_prod(a[i], b[j]) for all i and j.
+    """Computes the dot-product dot_prod(a[i], b[j]) for all i and j.
     :return: Matrix with res[i][j]  = dot_prod(a[i], b[j])
     """
     if not isinstance(a, torch.Tensor):
@@ -41,3 +49,218 @@ def dot_score(a: torch.Tensor, b: torch.Tensor):
         b = b.unsqueeze(0)
 
     return torch.mm(a, b.transpose(0, 1))
+
+
+# From https://github.com/beir-cellar/beir/blob/f062f038c4bfd19a8ca942a9910b1e0d218759d4/beir/retrieval/custom_metrics.py#L4
+def mrr(
+    qrels: dict[str, dict[str, int]],
+    results: dict[str, dict[str, float]],
+    k_values: List[int],
+) -> Tuple[Dict[str, float]]:
+    MRR = {}
+
+    for k in k_values:
+        MRR[f"MRR@{k}"] = 0.0
+
+    k_max, top_hits = max(k_values), {}
+    logging.info("\n")
+
+    for query_id, doc_scores in results.items():
+        top_hits[query_id] = sorted(
+            doc_scores.items(), key=lambda item: item[1], reverse=True
+        )[0:k_max]
+
+    for query_id in top_hits:
+        query_relevant_docs = set(
+            [doc_id for doc_id in qrels[query_id] if qrels[query_id][doc_id] > 0]
+        )
+        for k in k_values:
+            for rank, hit in enumerate(top_hits[query_id][0:k]):
+                if hit[0] in query_relevant_docs:
+                    MRR[f"MRR@{k}"] += 1.0 / (rank + 1)
+                    break
+
+    for k in k_values:
+        MRR[f"MRR@{k}"] = round(MRR[f"MRR@{k}"] / len(qrels), 5)
+        logging.info("MRR@{}: {:.4f}".format(k, MRR[f"MRR@{k}"]))
+
+    return MRR
+
+
+def recall_cap(
+    qrels: dict[str, dict[str, int]],
+    results: dict[str, dict[str, float]],
+    k_values: List[int],
+) -> Tuple[Dict[str, float]]:
+    capped_recall = {}
+
+    for k in k_values:
+        capped_recall[f"R_cap@{k}"] = 0.0
+
+    k_max = max(k_values)
+    logging.info("\n")
+
+    for query_id, doc_scores in results.items():
+        top_hits = sorted(doc_scores.items(), key=lambda item: item[1], reverse=True)[
+            0:k_max
+        ]
+        query_relevant_docs = [
+            doc_id for doc_id in qrels[query_id] if qrels[query_id][doc_id] > 0
+        ]
+        for k in k_values:
+            retrieved_docs = [
+                row[0] for row in top_hits[0:k] if qrels[query_id].get(row[0], 0) > 0
+            ]
+            denominator = min(len(query_relevant_docs), k)
+            capped_recall[f"R_cap@{k}"] += len(retrieved_docs) / denominator
+
+    for k in k_values:
+        capped_recall[f"R_cap@{k}"] = round(capped_recall[f"R_cap@{k}"] / len(qrels), 5)
+        logging.info("R_cap@{}: {:.4f}".format(k, capped_recall[f"R_cap@{k}"]))
+
+    return capped_recall
+
+
+def hole(
+    qrels: dict[str, dict[str, int]],
+    results: dict[str, dict[str, float]],
+    k_values: List[int],
+) -> Tuple[Dict[str, float]]:
+    Hole = {}
+
+    for k in k_values:
+        Hole[f"Hole@{k}"] = 0.0
+
+    annotated_corpus = set()
+    for _, docs in qrels.items():
+        for doc_id, score in docs.items():
+            annotated_corpus.add(doc_id)
+
+    k_max = max(k_values)
+    logging.info("\n")
+
+    for _, scores in results.items():
+        top_hits = sorted(scores.items(), key=lambda item: item[1], reverse=True)[
+            0:k_max
+        ]
+        for k in k_values:
+            hole_docs = [
+                row[0] for row in top_hits[0:k] if row[0] not in annotated_corpus
+            ]
+            Hole[f"Hole@{k}"] += len(hole_docs) / k
+
+    for k in k_values:
+        Hole[f"Hole@{k}"] = round(Hole[f"Hole@{k}"] / len(qrels), 5)
+        logging.info("Hole@{}: {:.4f}".format(k, Hole[f"Hole@{k}"]))
+
+    return Hole
+
+
+def top_k_accuracy(
+    qrels: dict[str, dict[str, int]],
+    results: dict[str, dict[str, float]],
+    k_values: List[int],
+) -> Tuple[Dict[str, float]]:
+    top_k_acc = {}
+
+    for k in k_values:
+        top_k_acc[f"Accuracy@{k}"] = 0.0
+
+    k_max, top_hits = max(k_values), {}
+    logging.info("\n")
+
+    for query_id, doc_scores in results.items():
+        top_hits[query_id] = [
+            item[0]
+            for item in sorted(
+                doc_scores.items(), key=lambda item: item[1], reverse=True
+            )[0:k_max]
+        ]
+
+    for query_id in top_hits:
+        query_relevant_docs = set(
+            [doc_id for doc_id in qrels[query_id] if qrels[query_id][doc_id] > 0]
+        )
+        for k in k_values:
+            for relevant_doc_id in query_relevant_docs:
+                if relevant_doc_id in top_hits[query_id][0:k]:
+                    top_k_acc[f"Accuracy@{k}"] += 1.0
+                    break
+
+    for k in k_values:
+        top_k_acc[f"Accuracy@{k}"] = round(top_k_acc[f"Accuracy@{k}"] / len(qrels), 5)
+        logging.info("Accuracy@{}: {:.4f}".format(k, top_k_acc[f"Accuracy@{k}"]))
+
+    return top_k_acc
+
+
+def get_rank_from_dict(
+    dict_of_results: dict[str, float], doc_id: str
+) -> Tuple[int, float]:
+    tuple_of_id_score = dict_of_results.items()
+    sorted_by_score = sorted(tuple_of_id_score, key=lambda x: x[1], reverse=True)
+    for i, (id, score) in enumerate(sorted_by_score):
+        if id == doc_id:
+            return i + 1, score
+
+    return len(sorted_by_score) + 1, 0
+
+
+def evaluate_change(
+    original_run: dict[str, dict[str, float]],
+    new_run: dict[str, dict[str, float]],
+    changed_qrels: dict[str, List[str]],
+) -> dict[str, float]:
+    changes = []
+    for qid in changed_qrels.keys():
+        original_qid_run = original_run[qid]
+        new_qid_run = new_run[qid]
+        for idx, changed_doc in enumerate(changed_qrels[qid]):
+            original_rank, original_score = get_rank_from_dict(
+                original_qid_run, changed_doc
+            )
+            new_rank, new_score = get_rank_from_dict(new_qid_run, changed_doc)
+            change = int(original_rank - new_rank)
+            changes.append(
+                {
+                    "qid": qid,
+                    "doc_id": changed_doc,
+                    "change": change,
+                    "relevance": 0,
+                    "og_rank": original_rank,
+                    "new_rank": new_rank,
+                    "og_score": original_score,
+                    "new_score": new_score,
+                }
+            )
+
+    # we now have a DF of [qid, doc_id, change] to run our calculations with
+    changes_df = pd.DataFrame(changes)
+    changes_df["p-MRR"] = changes_df.apply(lambda x: rank_score(x), axis=1)
+    qid_wise = changes_df.groupby("qid").agg({"p-MRR": "mean"})
+    return {
+        "p-MRR": qid_wise["p-MRR"].mean(),
+    }
+
+
+def rank_score(x: dict[str, float]) -> float:
+    if x["og_rank"] >= x["new_rank"]:
+        return ((1 / x["og_rank"]) / (1 / x["new_rank"])) - 1
+    else:
+        return 1 - ((1 / x["new_rank"]) / (1 / x["og_rank"]))
+
+
+# https://stackoverflow.com/a/62113293
+def download(url: str, fname: str):
+    resp = requests.get(url, stream=True)
+    total = int(resp.headers.get("content-length", 0))
+    with open(fname, "wb") as file, tqdm.tqdm(
+        desc=fname,
+        total=total,
+        unit="iB",
+        unit_scale=True,
+        unit_divisor=1024,
+    ) as bar:
+        for data in resp.iter_content(chunk_size=1024):
+            size = file.write(data)
+            bar.update(size)
