@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 
 
@@ -6,15 +8,12 @@ def _get_library_name() -> str:
 
 
 def _get_library_root_logger() -> logging.Logger:
-    """
-    Return the root logger of the library.
-    """
+    """Return the root logger of the library."""
     return logging.getLogger(_get_library_name())
 
 
 def enable_explicit_format() -> None:
-    """
-    Enable explicit formatting for every MTEB's logger. The explicit formatter is as follows:
+    """Enable explicit formatting for every MTEB's logger. The explicit formatter is as follows:
     ```
         [LEVELNAME|FILENAME|LINE NUMBER] TIME >> MESSAGE
     ```
@@ -23,5 +22,7 @@ def enable_explicit_format() -> None:
     handlers = _get_library_root_logger().handlers
 
     for handler in handlers:
-        formatter = logging.Formatter("[%(levelname)s|%(filename)s:%(lineno)s] %(asctime)s >> %(message)s")
+        formatter = logging.Formatter(
+            "[%(levelname)s|%(filename)s:%(lineno)s] %(asctime)s >> %(message)s"
+        )
         handler.setFormatter(formatter)
