@@ -1,25 +1,110 @@
-from .AmazonCounterfactualClassification import *
-from .AmazonPolarityClassification import *
-from .AmazonReviewsClassification import *
-from .AngryTweetsClassification import *
-from .Banking77Classification import *
-from .CMTEBClassification import *
-from .DalajClassification import *
-from .DanishPoliticalCommentsClassification import *
-from .DKHateClassification import *
-from .EmotionClassification import *
-from .ImdbClassification import *
-from .LccSentimentClassification import *
-from .MasakhaNEWSClassification import *
-from .MassiveIntentClassification import *
-from .MassiveScenarioClassification import *
-from .MTOPDomainClassification import *
-from .MTOPIntentClassification import *
-from .NordicLangClassification import *
-from .NoRecClassification import *
-from .NorwegianParliamentClassification import *
-from .PolishClassification import *
-from .ScalaClassification import *
-from .SweRecClassification import *
-from .ToxicConversationsClassification import *
-from .TweetSentimentExtractionClassification import *
+from __future__ import annotations
+
+from .ara.AJGT import *
+from .ara.HotelReviewSentimentClassification import *
+from .ara.RestaurantReviewSentimentClassification import *
+from .ara.TweetEmotionClassification import *
+from .ara.TweetSarcasmClassification import *
+from .bam.BambaraSentimentClassification import *
+from .ben.BengaliHateSpeechClassification import *
+from .ben.BengaliSentimentAnalysis import *
+from .bul.BulgarianSentimentClassification import *
+from .bul.BulgarianStoreReviewSentimentClassfication import *
+from .ces.CzechSubjectivityClassification import *
+from .dan.AngryTweetsClassification import *
+from .dan.DanishPoliticalCommentsClassification import *
+from .dan.DKHateClassification import *
+from .dan.LccSentimentClassification import *
+from .deu.GermanPoliticiansTwitterSentimentClassification import *
+from .ell.GreekLegalCodeClassification import *
+from .ell.GreekSentimentClassification import *
+from .eng.AmazonPolarityClassification import *
+from .eng.ArxivClassification import *
+from .eng.Banking77Classification import *
+from .eng.CanadaTaxCourtOutcomesLegalBenchClassification import *
+from .eng.ContractNLIConfidentialityOfAgreementLegalBenchClassification import *
+from .eng.ContractNLIExplicitIdentificationLegalBenchClassification import *
+from .eng.ContractNLIInclusionOfVerballyConveyedInformationLegalBenchClassification import *
+from .eng.ContractNLILimitedUseLegalBenchClassification import *
+from .eng.ContractNLINoLicensingLegalBenchClassification import *
+from .eng.ContractNLINoticeOnCompelledDisclosureLegalBenchClassification import *
+from .eng.ContractNLIPermissibleAcquirementOfSimilarInformationLegalBenchClassification import *
+from .eng.ContractNLIPermissibleCopyLegalBenchClassification import *
+from .eng.ContractNLIPermissibleDevelopmentOfSimilarInformationLegalBenchClassification import *
+from .eng.ContractNLIPermissiblePostAgreementPossessionLegalBenchClassification import *
+from .eng.ContractNLIReturnOfConfidentialInformationLegalBenchClassification import *
+from .eng.ContractNLISharingWithEmployeesLegalBenchClassification import *
+from .eng.DBpediaClassification import *
+from .eng.EmotionClassification import *
+from .eng.FinancialPhrasebankClassification import *
+from .eng.ImdbClassification import *
+from .eng.NewsClassification import *
+from .eng.PatentClassification import *
+from .eng.ToxicChatClassification import *
+from .eng.ToxicConversationsClassification import *
+from .eng.TweetSentimentExtractionClassification import *
+from .eng.YelpReviewFullClassification import *
+from .est.estonian_valence import *
+from .fas.PersianFoodSentimentClassification import *
+from .fil.FilipinoHateSpeechClassification import *
+from .fin.FinToxicityClassification import *
+from .fra.MovieReviewSentimentClassification import *
+from .guj.GujaratiNewsClassification import *
+from .hin.HindiDiscourseClassification import *
+from .hin.SentimentAnalysisHindi import *
+from .hrv.CroatianSentimentClassification import *
+from .ind.IndonesianIdClickbaitClassification import *
+from .ita.ItalianLinguistAcceptabilityClassification import *
+from .jpn.WRIMEClassification import *
+from .kan.KannadaNewsClassification import *
+from .kur.KurdishSentimentClassification import *
+from .mal.MalayalamNewsClassification import *
+from .mkd.MacedonianTweetSentimentClassification import *
+from .mlt.MalteseSentimentClassification import *
+from .multilingual.AmazonCounterfactualClassification import *
+from .multilingual.AmazonReviewsClassification import *
+from .multilingual.IndicLangClassification import *
+from .multilingual.IndicSentimentClassification import *
+from .multilingual.MasakhaNEWSClassification import *
+from .multilingual.MassiveIntentClassification import *
+from .multilingual.MassiveScenarioClassification import *
+from .multilingual.MTOPDomainClassification import *
+from .multilingual.MTOPIntentClassification import *
+from .multilingual.MultiHateClassification import *
+from .multilingual.NordicLangClassification import *
+from .multilingual.ScalaClassification import *
+from .multilingual.SIB200Classification import *
+from .multilingual.SouthAfricanLangClassification import *
+from .multilingual.TweetSentimentClassification import *
+from .nep.NepaliNewsClassification import *
+from .nld.DutchBookReviewSentimentClassification import *
+from .nob.NoRecClassification import *
+from .nob.NorwegianParliamentClassification import *
+from .ory.OdiaNewsClassification import *
+from .pan.PunjabiNewsClassification import *
+from .pol.PolishClassification import *
+from .por.HateSpeechPortugueseClassification import *
+from .ron.RomanianSentimentClassification import *
+from .san.SanskritShlokasClassification import *
+from .sin.SinhalaNewsClassification import *
+from .sin.SinhalaNewsSourceClassification import *
+from .slk.SlovakSentimentClassification import *
+from .slv.FrenkSlClassification import *
+from .spa.SpanishSentimentClassification import *
+from .ssw.SiswatiNewsClassification import *
+from .swe.DalajClassification import *
+from .swe.SweRecClassification import *
+from .tam.TamilNewsClassification import *
+from .tel.TeluguAndhraJyotiNewsClassification import *
+from .tha.wisesight_sentiment_classification import *
+from .tur.TurkishMovieSentimentClassification import *
+from .tur.TurkishProductSentimentClassification import *
+from .uig.UyghurSentimentClassification import *
+from .ukr.UkrFormalityClassification import *
+from .urd.UrduRomanSentimentClassification import *
+from .vie.VieStudentFeedbackClassification import *
+from .zho.CMTEBClassification import *
+from .zho.YueOpenriceReviewClassification import (
+    YueOpenriceReviewClassification,  # noqa: F401
+)
+from .zul.IsiZuluNewsClassification import *
