@@ -1,7 +1,14 @@
-from .CMTEBPairClassification import *
-from .OpusparcusPC import *
-from .PolishPC import *
-from .PawsX import *
-from .SprintDuplicateQuestionsPC import *
-from .TwitterSemEval2015PC import *
-from .TwitterURLCorpusPC import *
+from __future__ import annotations
+
+from .deu.FalseFriendsDeEnPC import *
+from .eng.CitationPredictionLegalBenchPC import *
+from .eng.ConsumerContractsQALegalBenchPC import *
+from .eng.SprintDuplicateQuestionsPC import *
+from .eng.TwitterSemEval2015PC import *
+from .eng.TwitterURLCorpusPC import *
+from .hye.ArmenianParaphrasePC import *
+from .multilingual.OpusparcusPC import *
+from .multilingual.PawsX import *
+from .multilingual.XNLI import *
+from .pol.PolishPC import *
+from .zho.CMTEBPairClassification import *
