@@ -1,16 +1,24 @@
-from .BiossesSTS import *
-from .CMTEBSTS import *
-from .GermanSTSBenchmarkSTS import *
-from .PolishSTS import *
-from .SickFrSTS import *
-from .SickrSTS import *
-from .STS12STS import *
-from .STS13STS import *
-from .STS14STS import *
-from .STS15STS import *
-from .STS16STS import *
-from .STS17CrosslingualSTS import *
-from .STS22CrosslingualSTS import *
-from .STSBenchmarkMultilingualSTS import *
-from .STSBenchmarkSTS import *
-from .STSES import *
+from __future__ import annotations
+
+from .deu.GermanSTSBenchmarkSTS import *
+from .eng.BiossesSTS import *
+from .eng.SickrSTS import *
+from .eng.STS12STS import *
+from .eng.STS13STS import *
+from .eng.STS14STS import *
+from .eng.STS15STS import *
+from .eng.STS16STS import *
+from .eng.STSBenchmarkSTS import *
+from .fin.FinParaSTS import *
+from .fra.SickFrSTS import *
+from .jpn.JSTS import *
+from .kor.KlueSTS import *
+from .kor.KorSTS import *
+from .multilingual.IndicCrosslingualSTS import *
+from .multilingual.STS17CrosslingualSTS import *
+from .multilingual.STS22CrosslingualSTS import *
+from .multilingual.STSBenchmarkMultilingualSTS import *
+from .pol.PolishSTS import *
+from .ron.RonSTS import *
+from .spa.STSES import *
+from .zho.CMTEBSTS import *
